+++ README.md
@@ -1,1 +1,53 @@
 # octo-squirrel
+
+For study purposes only.
+
+## Feature
+### Transmission
+|     | Shadowsocks | VMess |
+|:----|:-----------:|:-----:|
+| TCP |      ✔      |   ✔   |
+
+### Ciphers
+|                   | Shadowsocks |  VMess  |
+|:------------------|:-----------:|:-------:|
+| aes-128-gcm       |     `C`     |   `C`   |
+| aes-256-gcm       |     `C`     |         |
+| chacha20-poly1305 |     `C`     |   `C`   |
+
+`C` for client
+
+## How to run
+1. put config.json file before running
+
+    ```json
+    {
+        "servers":
+        [
+            {
+                "cipher": "{cipher}",
+                "password": "{password}",
+                "port": "{port}",
+                "protocol": "{protocol}",
+                "networks":
+                [
+                    "{networks}"
+                ]
+            }
+        ]
+    }
+    ```
+
+    > protocol: "shadowsocks" | "vmess"
+
+    > cipher: see Ciphers
+
+    > networks: see Transmission
+
+2. running command
+
+    * Windows
+
+    ```cmd
+    octo-squirrel-client.exe {config.json file path}
+    ```
